Rename package and appId (#1)
This allows us to re-publish the app as a new application.
