+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "29.0.2"
 
     defaultConfig {
-        applicationId "ch.fixme.status"
+        applicationId "io.spaceapi.myhackerspace"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 20
