+++ app/src/main/java/io/spaceapi/myhackerspace/Main.java
@@ -4,7 +4,7 @@
  * Licensed under GNU's GPL 3, see README
  */
 
-package ch.fixme.status;
+package io.spaceapi.myhackerspace;
 
 import android.app.Activity;
 import android.app.AlertDialog;
